requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
aiohttp==3.9.5
asyncpg==0.29.0
python-dotenv==1.0.0
httpx>=0.27.0
ultralytics>=8.3.0
torch>=2.3.0
torchvision>=0.18.0
timm>=1.0.0
streamlit==1.38.0
streamlit-folium==0.22.0
reportlab==4.2.5
folium==0.17.0
plotly==5.24.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
scikit-learn==1.5.0
requests==2.32.0
